fix: 전시회 생성페이지 수정
